More normalization
A copy of the full text of expressions is not stored anymore. This
reduces the size of the DB by about 15% and shaves a similar amount off
the training time.

Due to the above, the token separator is no longer needed. It has been
removed.

I moved most of the learning/replying logic into the storage backend so
we can reduce DB lookups without leaking some implementation details
(we cache some token ids now, etc).

Moved repeat_limit into the Storage role for now.