Ignore old Hailo modules in ->plugins that will fail to load due to changes in the API
Here's a bug report to Module::Pluggable that I didn't submit. I was
going to propose an option to M::P to only load modules from an @INC
directory that was under the directory which ->plugins was called
in. I.e. if Hailo.pm was in /opt/perl don't load from /opt/old-perl/.

However that won't work since users can install modules with
CPAN/CPANPLUS anywhere in @INC they want, and furthermore those tools
don't uninstall old code which could screw with our loading of new
modules in the same directories.

The only non-hardcoded solution to this that I can think of is to find
out the full path of each module we're about to load and check if its
$VERSION matches ours, an if not carry on.

But I don't feel like implementing my own require() so this patch will
do for the time being.

Here's the bug report:

I'm not sure I understand Module::Pluggable to suggest any particular
solution to the problem described below but I'd like to note it for
the record.

I have an application on CPAN (http://github.com/hinrik/hailo) that
does fuzzy loading of Storage backends based on user input. If you
tell it to load 'sqlite' it will do a regex search of
C<$self->plugins> for things that match C</sqlite/>, i.e. like this:

    # Be fuzzy about includes, e.g. DBD::SQLite or SQLite or sqlite will go
    my $pkg = first { / $type : .* : $class /ix }
              sort { length $a <=> length $b }
              $self->plugins;

The problem is that we recently moved some modules around so that in
earlier versions the SQLite module was called
C<Hailo::Storage::SQLite> but it's now called
C<Hailo::Storage::DBD::SQLite>. If I run tests installing say 0.10 on
a machine that previously had 0.05 installed they will fail since
we'll end up loading the old module from the filesystem instead of the
new one.

This is how I'm working around this:

    use Module::Pluggable (
        search_path => [ map { "Hailo::$_" } qw(Storage Tokenizer UI) ],
        except      => [
            qr[Mixin],

            # If an old version of Hailo is already istalled these modules
            # may be lying around. Ignore them manually; and make sure to
            # update this list if we move things around again.
            map( { qq[Hailo::Storage::$_] } qw(SQL SQLite Pg mysql)),
        ],
    );

Of course I'll have to manually update this blacklist in the future if
we end up moving anything else around; which is non-optimal.