Store store whitespace information with tokens
I made big changes to the NonWhitespace tokenizer and the SQLite backend
so that a normal/prefix/infix/postfix attribute is stored in the DB
along with the token in question. This only adds 4MB to the DB, and
allows us to store tokens in a more sane manner. E.g. when replying to
"foo", we there will only be one "foo" token to look up, not "foo;",
"foo:", "foo!", etc, so we will have a wider range of possible replies
to generate.