Hailo's hybrid file brain + in_memory SQLite mode now works, test it
It was broken because of the circular reference issue solved in "Don't
pass a copy of $self to slave classes to avoid circular refs".

The previous test worked by learning "foo bar baz", then checking the
disk size and learning "foo bar baz" again, then checking if the size
of the database changed on disk. The problem was that SQLite would
only change the values of integer fields in the database, which are
constant size.

That's fixed, and the test is now greatly expanded to test in-memory
hybrid mode.