Don't pass a copy of $self to slave classes to avoid circular refs
This is a very dirty hack that I'm using to get around Perl's
reference-based garbage collector. It's *UGLY*. What I wished I could
do instead was:

    commit 1c8c7d96e95d056921547df2c9286fa455c2a4df
    Author: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
    Date:   Sun Jul 18 20:58:27 2010 +0000

        What I wished I could do instead

    diff --git a/lib/Hailo.pm b/lib/Hailo.pm
    index 991ef81..9fadace 100644
    --- a/lib/Hailo.pm
    +++ b/lib/Hailo.pm
    @@ -5,6 +5,7 @@ use autodie qw(open close);
     use Any::Moose;
     use Any::Moose 'X::StrictConstructor';
     use List::Util qw(first);
    +use Scalar::Util qw(weaken);
     use namespace::clean -except => 'meta';

     use constant PLUGINS => [ qw[
    @@ -122,6 +123,8 @@ for my $k (keys %has) {
         no strict 'refs';
         *{"_build__${k}"} = sub {
             my ($self) = @_;
    +        my $s = $self;
    +        weaken($s);
             my $obj = $self->_new_class(
                 $name,
                 $self->$method_class,
    @@ -135,7 +138,7 @@ for my $k (keys %has) {
                      : ()),
                     (($k ~~ [ qw< storage > ] and defined $self->brain)
                      ? (
    -                     hailo => $self,
    +                     hailo => $s,
                          brain => $self->brain
                      )
                      : ()),

But unfortunately that doesn't work for some reason. Probably because
there's some reference to $self still around that I'm not catching
(maybe Moose hangs on to one?).

Anyway, meanwhile this dirty hack works, and passing the whole object
was always conceptually ugly anyway, you might even say that this new
way of only passing an anonymous sub interface is pretty. Yeah, I
don't believe it either.

Here's a writeup I mode for a Gist about the issue, for reference:

Disk -> in-memory -> disk mode doesn't work because hailo doesn't get
DEMOLISHED at the right time, which happens because I'm passing a
reference of $self to class that Hailo instansiates, which creates a
circular reference.

I'm passing the reference because the storage class needs to ask
things of Hailo which might have changed since the Storage class was
constructed, so they can't be passed at compile time.

In addition, the storage class needs to set things in the Hailo class
or its siblings (the Engine) if e.g. the order is changed:

    ack --nocolor --nogroup  -- '->hailo'
    Hailo/Storage.pm:143:        if ($self->hailo->_custom_order) {
    Hailo/Storage.pm:159:        $self->hailo->order($db_order);
    Hailo/Storage.pm:160:        $self->hailo->_engine->order($db_order);
    Hailo/Storage.pm:177:        if ($self->hailo->_custom_tokenizer_class) {
    Hailo/Storage.pm:195:        $self->hailo->tokenizer_class($db_tokenizer_class);

Refcount GC failing:

    $ perl -Ilib -MHailo -E '{ my $m = Hailo->new( brain => ":memory"); $m->learn("hello there man"); } warn "now here"'
    now here at -e line 1.
    Demolishing Hailo=HASH(0x798750) at lib/Hailo.pm line 315 during global destruction.

and:

    v hailo (master) $ perl -Ilib -MHailo -E '{ my $m = Hailo->new( brain => ":memory"); $m->learn("hello there man"); } warn "now here"'
    Demolishing Hailo=HASH(0x1a76750) at lib/Hailo.pm line 312.
    now here at -e line 1.